+++ Kernel/Config/Defaults.pm
@@ -904,20 +904,6 @@ sub LoadDefaults {
 #    $Self->{'MIME-Viewer'}->{'text/xml'} = $Self->{Home}.'/scripts/tools/xml2html.pl';
 
     # --------------------------------------------------- #
-    # SpellChecker
-    # --------------------------------------------------- #
-    # (If ispell or aspell is available, then we will provide a spelling
-    # checker.)
-    #    $Self->{SpellChecker} = 0;
-    $Self->{SpellChecker}            = 0;
-    $Self->{SpellCheckerBin}         = '/usr/bin/ispell';
-    $Self->{SpellCheckerDictDefault} = 'english';
-
-    # SpellCheckerIgnore
-    # (A list of ignored words.)
-    $Self->{SpellCheckerIgnore} = [ 'www', 'webmail', 'https', 'http', 'html', 'rfc' ];
-
-    # --------------------------------------------------- #
     # directories                                         #
     # --------------------------------------------------- #
     # root directory
@@ -1163,26 +1149,6 @@ sub LoadDefaults {
         'PasswordRegExp'                    => '',
         'Prio'                              => '0500',
         'Desc'                              => 'Set a new password by filling in your current password and a new one.',
-    };
-    $Self->{PreferencesGroups}->{SpellDict} = {
-        Module => 'Kernel::Output::HTML::Preferences::Generic',
-        PreferenceGroup => 'Miscellaneous',
-        Label  => 'Spelling Dictionary',
-        Desc   => 'Select your default spelling dictionary.',
-        Data   => {
-
-            # installed dict catalog (check your insalled catalogues, e. g. deutsch -=> german!)
-            # dict => frontend (ispell)
-            'english' => 'English',
-            'deutsch' => 'Deutsch',
-
-            # dict => frontend (aspell)
-#            'english' => 'English',
-#            'german'  => 'Deutsch',
-        },
-        PrefKey => 'UserSpellDict',
-        Prio    => 5000,
-        Active  => 1,
     };
     $Self->{PreferencesGroups}->{Comment} = {
         'Active'  => '0',