+++ Kernel/Modules/AgentTicketEmail.pm
@@ -2683,32 +2683,6 @@ sub _MaskEmailNew {
         );
     }
 
-    my $ShownOptionsBlock;
-
-    # show spell check
-    if ( $LayoutObject->{BrowserSpellChecker} ) {
-
-        # check if need to call Options block
-        if ( !$ShownOptionsBlock ) {
-            $LayoutObject->Block(
-                Name => 'TicketOptions',
-                Data => {
-                    %Param,
-                },
-            );
-
-            # set flag to "true" in order to prevent calling the Options block again
-            $ShownOptionsBlock = 1;
-        }
-
-        $LayoutObject->Block(
-            Name => 'SpellCheck',
-            Data => {
-                %Param,
-            },
-        );
-    }
-
     # Show the customer user address book if the module is registered and java script support is available.
     if (
         $ConfigObject->Get('Frontend::Module')->{AgentCustomerUserAddressBook}
@@ -2723,6 +2697,9 @@ sub _MaskEmailNew {
         Action => 'AdminCustomerUser',
         Type   => 'rw',
     );
+
+    my $ShownOptionsBlock;
+
     if ($OptionCustomer) {
 
         # check if need to call Options block