+++ Kernel/Modules/AgentTicketMove.pm
@@ -1464,18 +1464,6 @@ sub AgentMove {
             );
         }
 
-        # show spell check
-        if ( $LayoutObject->{BrowserSpellChecker} ) {
-            $LayoutObject->Block(
-                Name => 'TicketOptions',
-                Data => {},
-            );
-            $LayoutObject->Block(
-                Name => 'SpellCheck',
-                Data => {},
-            );
-        }
-
         # show attachments
         ATTACHMENT:
         for my $Attachment ( @{ $Param{Attachments} } ) {