+++ var/httpd/htdocs/js/Core.Agent.TicketAction.js
@@ -40,22 +40,6 @@ Core.Agent.TicketAction = (function (TargetNS) {
 
     /**
      * @private
-     * @name OpenSpellChecker
-     * @memberof Core.Agent.TicketAction
-     * @function
-     * @description
-     *      Open the spellchecker screen.
-     */
-    function OpenSpellChecker() {
-        var SpellCheckIFrame, SpellCheckIFrameURL;
-        SpellCheckIFrameURL = Core.Config.Get('CGIHandle') + '?Action=AgentSpelling;Field=RichText;Body=' + encodeURIComponent($('#RichText').val());
-        SpellCheckIFrameURL += SerializeData(Core.App.GetSessionInformation());
-        SpellCheckIFrame = '<iframe class="TextOption SpellCheck" src="' + SpellCheckIFrameURL + '"></iframe>';
-        Core.UI.Dialog.ShowContentDialog(SpellCheckIFrame, '', '10px', 'Center', true);
-    }
-
-    /**
-     * @private
      * @name OpenCustomerUserAddressBook
      * @memberof Core.Agent.TicketAction
      * @function
@@ -188,15 +172,6 @@ Core.Agent.TicketAction = (function (TargetNS) {
      */
     TargetNS.Init = function () {
 
-        // Initialize spell check functionality
-        TargetNS.InitSpellCheck();
-
-        // Register event for spell checker dialog
-        $('#OptionSpellCheck').on('click', function () {
-            OpenSpellChecker();
-            return false;
-        });
-
         $('.OptionCustomerUserAddressBook').on('click', function () {
             OpenCustomerUserAddressBook($(this).data('recipient-field'));
             return false;
@@ -216,42 +191,6 @@ Core.Agent.TicketAction = (function (TargetNS) {
 
         // Deactivate the fields in the template.
         Core.UI.InputFields.Deactivate($('#TemplateSelectionCustomerID'));
-
-        // check if spell check is being used
-        if (parseInt(Core.Config.Get('SpellChecker'), 10) === 1 && parseInt(Core.Config.Get('NeedSpellCheck'), 10) === 1) {
-
-            Core.Config.Set('TextIsSpellChecked', false);
-            $('#RichTextField, .RichTextField').on('click', '.cke_button__spellcheck', function() {
-                Core.Config.Set('TextIsSpellChecked', true);
-            });
-            $('#OptionSpellCheck').on('click', function() {
-                Core.Config.Set('TextIsSpellChecked', true);
-            });
-
-            if (parseInt(Core.Config.Get('RichTextSet'), 10) === 0) {
-                $('#RichTextField, .RichTextField').on('change', '#RichText', function() {
-                    Core.Config.Set('TextIsSpellChecked', false);
-                });
-            }
-
-            Core.Form.Validate.SetSubmitFunction($('form[name=compose]'), function() {
-                if ($('#RichText').val() && !$('#RichText').hasClass('ValidationIgnore') && !Core.Config.Get('TextIsSpellChecked')) {
-                    Core.App.Publish('Event.Agent.TicketAction.NeedSpellCheck', [$('#RichText')]);
-                    Core.UI.Dialog.ShowContentDialog('<p>' + Core.Language.Translate('Please perform a spell check on the the text first.') + '</p>', '', '150px', 'Center', true, [
-                        {
-                            Label: '<span>' + Core.Language.Translate('Close this dialog') + '</span>',
-                            Function: function () {
-                                Core.UI.Dialog.CloseDialog($('.Dialog:visible'));
-                                Core.Form.EnableForm($('#RichText').closest('form'));
-                            },
-                            Class: 'Primary CallForAction'
-                        }
-                    ]);
-                    return false;
-                }
-                $('#RichText').closest('form').get(0).submit();
-            });
-        }
 
         // Subscribe to the reloading of the CustomerInfo box to
         // specially mark the primary customer
@@ -289,60 +228,6 @@ Core.Agent.TicketAction = (function (TargetNS) {
             if ($WidgetElement.hasClass('Expanded')) {
                 $('#CreateArticle').prop('checked', true);
             }
-        });
-
-        // Subscribe to NeedSpellCheck event to open RTE widget if collapsed, if spellcheck is needed on submit
-        Core.App.Subscribe('Event.Agent.TicketAction.NeedSpellCheck', function ($TextElement) {
-            var $Widget = $TextElement.closest('div.WidgetSimple');
-
-            if ($Widget.attr('id') !== 'WidgetArticle' || $Widget.hasClass('Expanded')) {
-                return;
-            }
-
-            $Widget.find('div.WidgetAction.Toggle > a').trigger('click');
-        });
-
-    };
-
-    /**
-     * @name InitSpellCheck
-     * @memberof Core.Agent.TicketAction
-     * @function
-     * @returns {Boolean} Returns false, if the current action is not the spellchecker iframe
-     * @description
-     *      This function initializes the necessary stuff for spell check link  in TicketAction screens.
-     */
-    TargetNS.InitSpellCheck = function () {
-
-        if (Core.Config.Get('Action') !== 'AgentSpelling') {
-            return false;
-        }
-
-        // Register onchange event for dropdown and input field to change the radiobutton
-        $('#SpellCheck select, #SpellCheck input[type="text"]').on('change', function () {
-            var $Row = $(this).closest('tr'),
-                RowCount = parseInt($Row.attr('id').replace(/Row/, ''), 10);
-            $Row.find('input[type="radio"][id=ChangeWord' + RowCount + ']').prop('checked', true);
-        });
-
-        // Register Apply button event
-        $('#Apply').on('click', function () {
-            // Update ticket action popup fields
-            var FieldName = $('#Field').val(),
-                $Body = $('#' + FieldName, parent.document);
-
-            $Body.val($('#Body').val());
-
-            // Because we are in an iframe, we need to call the parent frames javascript function
-            // with a jQuery object which is in the parent frames context
-            parent.Core.UI.Dialog.CloseDialog($('.Dialog', parent.document));
-        });
-
-        // Register Cancel button event
-        $('#Cancel').on('click', function () {
-            // Because we are in an iframe, we need to call the parent frames javascript function
-            // with a jQuery object which is in the parent frames context
-            parent.Core.UI.Dialog.CloseDialog($('.Dialog', parent.document));
         });
     };